Close #87 - Add types package and move numeric, strings and network to it
