+++ modules/refined4s-core/shared/src/main/scala/refined4s/types/network.scala
@@ -1,4 +1,6 @@
-package refined4s
+package refined4s.types
+
+import refined4s.*
 
 import scala.quoted.*
 import scala.util.control.NonFatal
