+++ modules/refined4s-core/shared/src/main/scala/refined4s/types/numeric.scala
@@ -1,5 +1,6 @@
-package refined4s
+package refined4s.types
 
+import refined4s.*
 import refined4s.internal.numericTools
 
 import numeric.*
